feat: 신고 기능 추가하기
- 신고 기능 추가하기

+++ MATFLIX/src/main/java/com/boot/dto/ReportDisplayDTO.java
@@ -10,11 +10,11 @@ import lombok.NoArgsConstructor;
 @AllArgsConstructor
 @NoArgsConstructor
 public class ReportDisplayDTO {
-	private Long report_id; // 신고 PK
-	private Integer reporter_mf_no; // 신고자 mf_no
+	private int report_id; // 신고 PK
+	private int reporter_mf_no; // 신고자 mf_no
 	private String target_type; // USER / BOARD / COMMENT / RECIPE
-	private Long target_id; // 대상 PK
-	private Integer target_owner_mf_no; // 대상 작성자 mf_no
+	private int target_id; // 대상 PK
+	private int target_owner_mf_no; // 대상 작성자 mf_no
 	private String report_title;
 	private String report_reason; // 성적 / 욕설 / 비하 / 차별 / 스팸 등
 	private String report_detail;
